Fix formatting and linting. (#8)
* Fix formatting and linting.

In first place, let me apologize. I needed to fix the Finnish formatting
but linting got into my editor's way so this is the complete list of
changes:

- Fix formatting for a few languages. Some languages include a dot in
the short format and in the data we get that dot surrounded by single
quotes. This is the main reason for this PR.
- Make the project pass its linting rules.
- Make CI check linting.
- Fix path in tslint.json to properly exclude test/locale-data.ts
- Add test/locale-data.ts to .prettierignore

* Fix versions of node to use

Jest is having a problem with latest unstable node (11.11.0) https://github.com/facebook/create-react-app/issues/6591 so trying to set specific supported versions to fix travis

* Fix specifying versions